backend/requirements.txt
# Core dependencies
numpy>=1.22.0
pandas>=1.4.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Deep Learning
torch>=1.13.0
torchvision>=0.14.0

# Data processing
pyyaml>=6.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Project utilities
tqdm>=4.64.0
